requirements.txt
Flask==2.0.3
Flask-PyMongo==2.2.0
flask-cors==3.0.10
